Strictly validate action names
This is not required by the spec, but action names are intended to be
IDs and it is possible that the notification daemon was not tested with
strange characters in these IDs.  This could cause e.g. XML or SQL
injection if the daemon implementation is buggy.  Guard against this by
strictly validating the IDs.  Also block overly long action names in
case there is a buffer overflow in some notification daemon.