Automatically restart notification agent on failure
Set automatic restart so that after user logout in GUI domain the
service will attempt to re-connect later. Set restart delay to 1s to not
spam with attempts too often (all VMs using this service will try to
re-connect...). But do not restart if service call is denied, or when
service in GUI domain is not installed.