Slow down qubes-notification-agent.service restart
When user is not logged into dom0 (or GUI domain) yet, slow down restart
to avoid hitting restart limit, but still eventually connect to the
service.

This is not ideal solution to the problem, but an acceptable workaround,
with less risks of hitting some corner case. See linked issue for more
details.

Fixes QubesOS/qubes-issues#10137